+++ packages/js-dpp/CHANGELOG.md
@@ -1,3 +1,21 @@
+## [0.19.2](https://github.com/dashevo/js-dpp/compare/v0.19.1...v0.19.2) (2021-05-20)
+
+
+### Bug Fixes
+
+* Cbor not decoding buffers properly in browsers ([#306](https://github.com/dashevo/js-dpp/issues/308))
+
+
+
+## [0.19.1](https://github.com/dashevo/js-dpp/compare/v0.19.0...v0.19.1) (2021-05-04)
+
+
+### Bug Fixes
+
+* `topLevelIdentity.getId` is not a function ([#306](https://github.com/dashevo/js-dpp/issues/306))
+
+
+
 # [0.19.0](https://github.com/dashevo/js-dpp/compare/v0.18.0...v0.19.0) (2021-04-30)